requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
Pillow==10.2.0
google-generativeai==0.3.2